Reesctructuro las carpetas para los views y tests
