+++ app/web/tests/test_registro.py
