+++ app/web/views/administracion_views.py
@@ -1,0 +1,29 @@
+from django.shortcuts import render
+
+
+def administracion(request):
+    context = {
+        'title': 'Administrador'
+    }
+    return render(request, 'admin/administrador.html', context)
+
+
+def comercioAdmin(request):
+    context = {
+        'title': 'Comercio Admin'
+    }
+    return render(request, 'admin/comercios.html', context)
+
+
+def repartidorAdmin(request):
+    context = {
+        'title': 'Repartidor Admin'
+    }
+    return render(request, 'admin/repartidor.html', context)
+
+
+def clienteAdmin(request):
+    context = {
+        'title': 'Cliente Admin'
+    }
+    return render(request, 'admin/cliente.html', context)
