+++ app/web/views/registro_views.py
@@ -1,59 +1,15 @@
 from django.conf import settings
 from django.shortcuts import render
 from django.urls import reverse_lazy
-from django.http import HttpResponse
 from django.views.generic import FormView
-from django.http import HttpResponse, JsonResponse
+from django.http import JsonResponse
 from django.core.signing import Signer
 
 from django.utils import timezone
-
-from web.forms.loginForm import LoginForm
-from web.forms.contactoForm import ContactoForm
 from web.forms.registroForm import RegistroForm
 
 from django.core.mail import send_mail
 from web.models.usuario import Usuario
-
-
-def home(request):
-    loginForm = LoginForm()
-    contactoForm = ContactoForm()
-    registroForm = RegistroForm()
-    context = {
-        'loginForm': loginForm,
-        'contactoForm': contactoForm,
-        'registroForm': registroForm
-    }
-    return render(request,'web/home.html', context)
-
-
-def administracion(request):
-    context = {
-        'title': 'Administrador'
-    }
-    return render(request, 'admin/administrador.html', context)
-
-
-def comercioAdmin(request):
-    context = {
-        'title': 'Comercio Admin'
-    }
-    return render(request, 'admin/comercios.html', context)
-
-
-def repartidorAdmin(request):
-    context = {
-        'title': 'Repartidor Admin'
-    }
-    return render(request, 'admin/repartidor.html', context)
-
-
-def clienteAdmin(request):
-    context = {
-        'title': 'Cliente Admin'
-    }
-    return render(request, 'admin/cliente.html', context)
 
 
 def registroRealizado(request, email):
@@ -63,11 +19,6 @@ def registroRealizado(request, email):
     return render(request,'web/registro_realizado.html', context)
 
 
-########################
-#
-# Usuarios: Registro
-#
-########################
 
 class RegistroView(FormView):
     form_class = RegistroForm
@@ -88,7 +39,7 @@ class RegistroView(FormView):
         )
         return JsonResponse({'success': True, 'email': form.cleaned_data['email']})
 
-    def form_invalid(self, form) -> HttpResponse:
+    def form_invalid(self, form) -> JsonResponse:
         errors = form.errors
         return JsonResponse({'errors': errors}, status=400)
     
@@ -112,9 +63,11 @@ def activarCuenta(request, key):
         usuario.save()
         
         #TODO: crear variable de sesión
+        # para habilitar la cuenta tengo que completar los datos
+        # dependiendo el perfil del usuario
         if usuario.tipo.titulo == 'Cliente':
             context = {'tipo':'cliente'}
-            return render(request, 'admin/activar_cuenta.html',context)
+            return render(request, 'admin/habilitar_cuenta.html',context)
     else:
         #TODO: hacer pagina 404
         print('Redirijo a una pantalla de Error')
